+++ crates/sqlite-store/src/sql_error.rs
@@ -20,15 +20,13 @@ impl<T> SqlResultExt<T> for Result<T, rusqlite::Error> {
             | rusqlite::Error::InvalidParameterCount(..)
             | rusqlite::Error::QueryReturnedNoRows => StoreError::QueryError(value.to_string()),
             rusqlite::Error::SqliteFailure(error, _) => match error.code {
-                rusqlite::ErrorCode::DatabaseBusy => StoreError::DatabaseBusy(value.to_string()),
-                rusqlite::ErrorCode::DatabaseLocked => {
-                    StoreError::DatabaseLocked(value.to_string())
+                rusqlite::ErrorCode::DatabaseBusy | rusqlite::ErrorCode::DatabaseLocked => {
+                    StoreError::DatabaseTransientError(value.to_string())
                 },
-                rusqlite::ErrorCode::ConstraintViolation => {
-                    StoreError::ConstraintViolation(value.to_string())
-                },
-                rusqlite::ErrorCode::DatabaseCorrupt | rusqlite::ErrorCode::NotADatabase => {
-                    StoreError::DatabaseCorrupted(value.to_string())
+                rusqlite::ErrorCode::ConstraintViolation
+                | rusqlite::ErrorCode::DatabaseCorrupt
+                | rusqlite::ErrorCode::NotADatabase => {
+                    StoreError::DatabasePermanentError(value.to_string())
                 },
                 _ => StoreError::DatabaseError(value.to_string()),
             },